Add CRUD for the Country, add a GET URL to result models
Fix #12
Relates to #21, #14

Add CRUD methods for the Country including Add with POST, Update with PUT and Delete with DELETE. Also you can add a currency to an existing country with POST to the different url and remove it with DELETE.
Add an Organization repository.

Add the GetURL field to result models (country, currency). This field contains the URL to get a corresponding item.